move select methods to op folder
